🚚 Move file related fragments, mutations and queries to /documents
