+++ src/documents/fragments.js
@@ -1,0 +1,24 @@
+import gql from 'graphql-tag';
+
+export const FILE_FIELDS = gql`
+  fragment FileFields on FileNode {
+    id
+    kfId
+    name
+    description
+    fileType
+    downloadUrl
+  }
+`;
+
+export const VERSION_FIELDS = gql`
+  fragment VersionFields on VersionNode {
+    id
+    kfId
+    fileName
+    size
+    state
+    description
+    createdAt
+  }
+`;
