+++ src/state/fragments.js
@@ -65,29 +65,6 @@ export const PROJECT_FIELDS = gql`
   }
 `;
 
-export const FILE_FIELDS = gql`
-  fragment FileFields on FileNode {
-    id
-    kfId
-    name
-    description
-    fileType
-    downloadUrl
-  }
-`;
-
-export const VERSION_FIELDS = gql`
-  fragment VersionFields on VersionNode {
-    id
-    kfId
-    fileName
-    size
-    state
-    description
-    createdAt
-  }
-`;
-
 export const CREATOR_FIELDS = gql`
   fragment CreatorFields on UserNode {
     id
